Sort skipped extensions in test summary
Sort flags are the same as for $exts_to_test (flags: SORT_REGULAR) in
main():730.

There was also a slight chance to branch into get_summary() with
(show_ext_summary: true) and count(NULL) use:  Given the call of
find_files() in main() on line 635, when there are no extensions to
skip, then the global variable $exts_skipped is not converted to an
array ($global[] = ...), that errors on count() then.

Therefore, we cast the global variable $exts_skipped to array before
sorting and counting.