Untrack test generated .data file
The cve-2014-3538-mb.phpt test similar to cve-2014-3538.phpt removes
the .data file in --CLEAN-- by unlink().

However, in 3d3f11ede4c ("Fixed the UTF-8 and long path support in the
streams on Windows.", 2016-06-20) the temporary fixture file has been
committed.

As the file is generated this is counter-productive.  Furthermore,
when running the tests under version control given a build from checkout,
then the VCS marks the file as deleted.

Fix is to remove the file from version control:

    $ git rm --cached ext/fileinfo/tests/cve-2014-3538私はガラスを食べられます.data

To prevent similar mistakes, ignore those test .data files in git with the
binary attribute macro to levitate the known issue in diff fetching tools.